This covers an edge case where the user doesn't have qt installed (or otherwise has not compiled it on a first run with jasmine:headless), and are then greeted on the web with a NoRunnerError.
Well, since we don't need the headless runner to
actually work when we're running specs in the browser,
this hack will just rescue the exception and then
do a head-fake by creating the file that JHW is
actually looking for, instantiate the runner, and
then delete it.

We need an instance of the runner to get at the 
Jasmine config in the same way JHW is.