Fix broken spec.
The test suite for rails 4.0 using jasmine 2 was failing. This suite
tests that jasmine-rails was successfully mounted under /specs by
asserting that the stats banner is shown.

In the current version of jasmine 2, the class names that this banner is
found under have changed to .jasmine-bar.jasmine-passed (presumeably
from .bar.passed)